moved s3 functions to utils
